🐛 가격 입력 타입 Long 변경 및 Line 1줄로 고정

+++ core/common/src/main/java/com/devd/common/ui/register/PriceRegisterPopup.kt
@@ -129,8 +129,8 @@ fun PriceRegisterPopup(
 fun PriceRegisterScreen(
     uiState: PriceRegisterUiState,
     updateProductName: (String) -> Unit,
-    updateRegularPrice: (Int) -> Unit,
-    updatePurchasePrice: (Int) -> Unit,
+    updateRegularPrice: (Long) -> Unit,
+    updatePurchasePrice: (Long) -> Unit,
     updateSelectStore: (Int) -> Unit,
     updateQuantity: (Long) -> Unit,
     updateSelectQuantity: (Int) -> Unit,
